+++ packages/realm-network-transport/src/NetworkTransport.ts
@@ -126,6 +126,12 @@ export class DefaultNetworkTransport implements NetworkTransport {
         this.fetch(request)
             .then(async response => {
                 const decodedBody = await response.text();
+                if (response.status >= 400) {
+                    throw {
+                        statusCode: response.status,
+                        errorMessage: decodedBody,
+                    };
+                }
                 // Pull out the headers of the response
                 const responseHeaders: Headers = {};
                 response.headers.forEach((value, key) => {