+++ CHANGELOG.md
@@ -25,7 +25,7 @@ NOTE 1: Requires MacOS v10.14 (Mojave) to build.
 
 ### Internal
 * Updated Realm Core from ??? to Realm Core v10.0.0-alpha.8.
-* Updated Realm Sync from ??? to Realm Sync v10.0.0-alpha.12.
+* Updated Realm Sync from ??? to Realm Sync v10.0.0-alpha.14.
 * Updated Realm Object Store to commit ec6ffd5cc149e403d252e9d615b4e6dab572fa89.
 * Created a package named "realm-app-importer", to be used by integration tests (ideally by other SDKs too).